requirements.txt
asgiref==3.8.1
Django==4.2.25
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0
requests==2.31.0
paypalrestsdk==1.13.1
PyJWT==2.10.1
